requirements.txt
streamlit>=1.32.0
sentence-transformers>=2.5.0
faiss-cpu>=1.8.0
pymupdf>=1.23.0
pytesseract>=0.3.10
Pillow>=10.2.0
numpy>=1.26.0
pandas>=2.2.0
torch>=2.2.0
# Phase 2
watchdog>=4.0.0
beautifulsoup4>=4.12.3
requests>=2.31.0
sumy>=0.11.0
nltk>=3.8.1
lxml>=5.1.0
pyyaml>=6.0
